Created Collisions and Implemented Changing movement direction
Created a new game object for a platform on which the player can stand.

Created 2D collision boxes for the player and platform.

Created RigidBody2D component for the player

Created Scripts folder in the Assets folder. Created the PlayerController C# script in the Scripts folder.

Implemented left and right changing movement direction for the player by making the x axis Player object scale negative or positive in the C# script.