Created Tile Map Palette for the Ground and added Level End trigger
Created the Grid game object. Created the Ground tile map and created the ground Tile palette. Added all the components in the tile palette and in turn added them in the game screen to create the first level. Added the TileMap2D collider for the GroundTileMap.

Created a game object with collider to detect Level end when it collides with player.