Implemented Key Collectible and Score text
Created a new game object for Key as a prefab asset.

Created a box 2D collider for the key.

Defined the function to display a message in the PlayerController script which will be called when the Player collides with/picks up the Key.

Created TextMeshPro component as a part of the Game UI to display the Player Score.

Created a C# script ScoreController for the Score text component and defined function to increase score.

Created a C# script KeyController for the Key prefab to destroy Key object and call the function defined in PlayerController and ScoreController every time the Player collides/picks up the key.