Implemented Enemy and Player Death
Created new game object for Enemy. Created Box 2D Collider for the Enemy.

Created C# script EnemyController for Enemy.

Created Death Animation for the Player.

Defined function in the PlayerController Script to play the player death animation and print player death message which will be called from the EnemyController Script when the Player collides with Enemy.