Created Lobby Screen and Game Over Screen
Created Canvas to include Game Over Image and Restart Game buttons in the Start Scene. Added RestartGame function of the LevelRestartController as the function to be called when the Restart button is clicked.

Created a new scene for a Lobby Screen Added the Title text and button to play the game. Created C# script LobbyController and defined the Play Game function to load the Start scene. Added the Play Game function as the function  to be called when the Play button is clicked.

Made changes to the Enemies by removing the enemy movement scripts and adding the movement logic in the enemy controller itself.