Created Multiple Levels and added Level selection screen in the lobby
Created 2 more scenes for 2 additional levels.

Created the Game Over Screen as a prefab and added a button to it to quit back to Lobby/Main Menu screen. Created C# Script QuitGameController and defined function to load the lobby screen which gets executed when the Quit button is clicked on the Game Over screen.

Added a popup screen to allow player to select between the 3 levels in the Lobby Screen which appears after clicking the play button. Defined functions in the LobbyController script to load each level.