Implemented Locked and Unlocked levels
Created the C# script LevelStatus and defined an enum which stores the Level status (Locked, Unlocked, Completed).

Created C# script LevelManager to get and set the level status and to load the next level when current level is completed.

Defined onClick function in the LobbyController C# script for the buttons in the LevelSelection popup which checks for the status of the Level and loads the level if it is unlocked or completed.