Added background music, sound effects and added Background Image
Made changes to the Lobby screen and Level screens by adding background image and an image for the level selection popup.

Added sound to the game by using Unity's Audio Source component and creating a C# script called Sound Manager to play different sounds for different events in the game. Created Sound Manager object as Singleton in the Lobby Scene.