Created the design for Level 3
Added vertical moving platforms in Level 3. Created new game objects for the vertical moving platforms and created them as prefabs. Created C# scripts for moving up and moving down platforms and defined functions which make the platform move from top to bottom or vice versa only when play is standing on it.

Added Game End design for the end of this level.